Move GenerateCommandConfigurationValidator
